refactor(GetResult) : Rank 리팩토링
- 구하는 방법 배열을 사용해서 변경
- 상금 long 타입으로 변경

+++ src/main/java/lotto/dto/ResultLineDTO.java
@@ -2,14 +2,14 @@ package lotto.dto;
 
 public class ResultLineDTO {
     private final String label;
-    private final int prize;
+    private final long prize;
     private final int count;
 
     public String getLabel() {
         return label;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
 
@@ -17,7 +17,7 @@ public class ResultLineDTO {
         return count;
     }
 
-    public ResultLineDTO(String label, int prize, int count) {
+    public ResultLineDTO(String label, long prize, int count) {
         this.label = label;
         this.prize = prize;
         this.count = count;
